\name{mthesis-package}
\alias{mthesis-package}
\alias{mthesis}
\docType{package}
\title{
\packageTitle{mthesis}
}
\description{
\packageDescription{mthesis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mthesis}
\packageIndices{mthesis}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mthesis}

Maintainer: \packageMaintainer{mthesis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
